DDF-1565 Changed Managed Service Factories from Spring to Blueprint. Affected Packages : content-core-directorymonitor catalog-opensearch-source spatial-csw-source spatial-kml-transformer spatial-wfs-mapper spatial-wfs-1_0_0-source spatial-wfs-2_0_0-source platform-mime-core-configureableresolver platform-scheduler

+++ catalog/spatial/wfs/1.0.0/spatial-wfs-v1_0_0-source/src/test/java/org/codice/ddf/spatial/ogc/wfs/v1_0_0/catalog/source/TestWfsSource.java
@@ -1028,8 +1028,8 @@ public class TestWfsSource {
         source.setConnectionTimeout(10000);
         source.setReceiveTimeout(10000);
         // Perform test
-        assertEquals(source.getConnectionTimeout(), 10000);
-        assertEquals(source.getReceiveTimeout(), 10000);
+        assertEquals(source.getConnectionTimeout().intValue(), 10000);
+        assertEquals(source.getReceiveTimeout().intValue(), 10000);
     }
 
     private SourceResponse executeQuery(int startIndex, int pageSize)
